+++ packages/aim_cli/lib/aim_cli.dart
@@ -3,3 +3,4 @@ library;
 
 export 'src/commands/create_command.dart';
 export 'src/commands/dev_command.dart';
+export 'src/commands/build_command.dart';